pyproject.toml
[project]
name = "ai-engineering-lab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.68.0",
    "groq>=1.1.1",
    "ipykernel>=7.2.0",
    "nbconvert>=7.17.0",
    "openai>=2.29.0",
    "openpyxl>=3.1.0",
    "pandas>=2.2.0",
    "python-docx>=1.1.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
]

[tool.uv.workspace]
members = [
    "apps/api",
    "apps/chatbot_ui",
]
